Use new synchronization around fork().
When a process forks, all the memory is copied, but only the thread
executing the fork() system call is copied.

One consequence is that if a mutex is locked by a thread other than the
forking thread (or a condition variable is being waited on), in the
child process it will cause undefined behavior (and possibly a process
hang) when deleting the mutex or condition variable.

Since the thread holding the mutex or waiting on the condition variable
does not exist in the child, there's no way to resolve this in the
child.

The approach here is to substitute a non-{mutex, condition variable}
mechanism to synchronize (halt) the profiler thread across the fork()
call. This ensures that the mutex is unlocked in the child and no thread
is waiting on the condition variable.