Add the following:
* Random sample interval to reduce the total number of samples, but not
extend the normal sample interval. The result is like turning the
sampling on and off at random intervals for a significantly longer
period than the randomized sampling interval. This preserves small
sampling interval to sample short-lived work.

* Use a mmap region to track the stack frames.

* Add GC awareness to not sample during GC and to mark the sampled stack
frames if GC occurs after we release the GVL but before we are done
processing that sample and take the next sample.